backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
requests>=2.28
fpdf>=1.7
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
Werkzeug>=2.0
python-dotenv>=1.0